04_CHAINLIT/requirements.txt
agno==1.1.13
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.3.0
bidict==0.23.1
certifi==2025.1.31
chainlit==2.4.1
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
dataclasses-json==0.6.7
datasets==3.4.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.11
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.19.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
joblib==1.4.2
lancedb==0.21.1
Lazify==0.4.0
literalai==0.1.103
llama-index-agent-openai==0.4.6
llama-index-core==0.12.25
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.25
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.4.1
mdurl==0.1.2
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.66.3
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pgvector==0.4.0
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.24.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sse-starlette==2.2.1
starlette==0.41.3
syncer==2.0.3
tenacity==9.0.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.31.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
